changed all Doubles to bigDecimal

+++ sales-tax-rates/src/main/java/com/complyt/domain/RatesMetaData.java
@@ -2,6 +2,8 @@ package com.complyt.domain;
 
 import lombok.With;
 
+import java.math.BigDecimal;
+
 @With
-public record RatesMetaData(double cityDistrictRate, double countyDistrictRate) {
+public record RatesMetaData(BigDecimal cityDistrictRate, BigDecimal countyDistrictRate) {
 }
