+++ sales-tax-rates/src/main/java/com/complyt/v1/model/RatesMetaDataDto.java
@@ -2,6 +2,8 @@ package com.complyt.v1.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 @Schema(name = "RatesMetaData")
-public record RatesMetaDataDto(double cityDistrictRate, double countyDistrictRate) {
+public record RatesMetaDataDto(BigDecimal cityDistrictRate, BigDecimal countyDistrictRate) {
 }
