+++ sales-tax/src/main/java/com/complyt/business/transaction/items_amounts/AmountCalculator.java
@@ -2,6 +2,8 @@ package com.complyt.business.transaction.items_amounts;
 
 import lombok.NonNull;
 
+import java.math.BigDecimal;
+
 public interface AmountCalculator<T> {
-    float calculate(@NonNull T t);
+    BigDecimal calculate(@NonNull T t);
 }
