+++ sales-tax/src/main/java/com/complyt/v1/models/sales_tax/SalesTaxRatesDto.java
@@ -2,7 +2,9 @@ package com.complyt.v1.models.sales_tax;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 @Schema(name = "SalesTaxRates")
-public record SalesTaxRatesDto(double cityRate, double countyRate, double stateRate, double taxRate,
-                               double combinedDistrictRate, RatesMetaDataDto ratesMetaData) {
+public record SalesTaxRatesDto(BigDecimal cityRate, BigDecimal countyRate, BigDecimal stateRate, BigDecimal taxRate,
+                               BigDecimal combinedDistrictRate, RatesMetaDataDto ratesMetaData) {
 }
